User Online Status
Fixed:
- archived ads displaying on UI;
- render upgrade membership options in a separate blade, removed 403 error blade
- created Form Request for creating and updating ads
- fixed error where user could activate ads with no available slots
- CityFactory now uses cities-in_germany.csv as resource
- seed data added + Image seeder
- fixed filter bugs
- fixed bug where users could delete the default avatar logo
- fixed erratic behaviour of the profile sections dropdowns
- improved visibility of sections in ShowSections component
- Vue pagination and tooltip tweaks

+++ resources/js/components/LogoForm.vue
@@ -69,9 +69,11 @@ export default {
 		},
 
 		removeImage() {
-			axios.delete(`/myaccount/settings/${this.user.id}/logos/delete`)
-				.then(this.avatar = 'users/default.png')
-				.then(() => flash('Image deleted!'));
+			if (this.user.avatar !== 'users/default.png') {
+				axios.delete(`/myaccount/settings/${this.user.id}/logos/delete`)
+					.then(this.avatar = '/storage/users/default.png')
+					.then(() => flash('Image deleted!'));
+			}
 		}
 	}
 };
